+++ lib/rdkafka.rb
@@ -68,9 +68,12 @@ require "rdkafka/producer/partitions_count_cache"
 require "rdkafka/producer"
 require "rdkafka/producer/delivery_handle"
 require "rdkafka/producer/delivery_report"
+require "rdkafka/clients"
 
 # Main Rdkafka namespace of this gem
 module Rdkafka
 end
 
 Rdkafka::Bindings.rd_kafka_global_init
+
+at_exit { Rdkafka::Clients.close_all }